fix: async 설정 변경
