+++ src/main/java/com/woowa/woowakit/global/config/AsyncConfig.java
@@ -1,20 +1,24 @@
 package com.woowa.woowakit.global.config;
 
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.AsyncConfigurerSupport;
 import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 
-@EnableAsync
-@Configuration
-public class AsyncConfig {
+import java.util.concurrent.Executor;
 
-	@Bean(name = "StockProcessingService", destroyMethod = "shutdown")
-	public ThreadPoolTaskExecutor healthCheckRequesterTaskExecutor() {
+@Configuration
+@EnableAsync
+public class AsyncConfig extends AsyncConfigurerSupport {
+
+	@Override
+	public Executor getAsyncExecutor() {
 		ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
-		executor.setCorePoolSize(5);
-		executor.setMaxPoolSize(20);
-		executor.setQueueCapacity(Integer.MAX_VALUE);
+		executor.setCorePoolSize(20);
+		executor.setMaxPoolSize(30);
+		executor.setQueueCapacity(1000000);
+		executor.setThreadNamePrefix("DNDN-ASYNC-");
+		executor.initialize();
 		return executor;
 	}
 }
