+++ src/main/java/com/woowa/woowakit/global/config/BatchAsyncConfig.java
@@ -1,2 +1,18 @@
-package com.woowa.woowakit.global.config;public class BatchAsynConfig {
+package com.woowa.woowakit.global.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
+
+@Configuration
+public class BatchAsyncConfig {
+
+	@Bean(name = "StockProcessingService", destroyMethod = "shutdown")
+	public ThreadPoolTaskExecutor healthCheckRequesterTaskExecutor() {
+		ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
+		executor.setCorePoolSize(5);
+		executor.setMaxPoolSize(20);
+		executor.setQueueCapacity(Integer.MAX_VALUE);
+		return executor;
+	}
 }
